build-falter: patch imagebuilder json-info-generation
This introduces patching the json-info-generation in the imagebuilder.
That way, we will get a whole list of the built devices, instead of
only the last one. The adjusted script was provided by @mwarning42.
Many thanks to his really excellent work on the image selector and
the really fast help on that issue.

Signed-off-by: Martin Hübner <martin.hubner@web.de>