build_falter: select openwrt-version via freifunk-release
In the freifunk_release file of each feed there are variables which
define the OpenWrt-Version the feed is intended for. This commit
introduces changes to use that OpenWrt-Verison mentioned there.

The option `-v` now encodes, which falter-version to use. The version
maps directly to the package feed at the buildbot.

This fixes #63.

Signed-off-by: Martin Hübner <martin.hubner@web.de>