scripts: add a script to update the olsrd init scripts
use the upstream olsrd and olsrd6 init scripts.  The init scripts
delivered with 21.02.0 are problematic and the changes needed have
been incorperated upstream.  This script will execute when the
OPENWRT_BASE is set to "21.02.0".  For all other versions, the package
default will be used.

The relevant commit it to get the updated init scripts is:
1fcda9dfa893ed5b06ce71f36f454cfec11092e5

Signed-off-by: pmelange <isprotejesvalkata@gmail.com>