build_falter: implement reducing packagelists for 8MiB-Devices
Devices with 8MiB Flash ran out of memory when storing changes
on the configuration. This commit tackles that problem by
automatically removing some not crucial packages, if we detect
a device with 8MiB flash.

This is achieved by loading the OpenWrt Table of Hardware into
a sqlite-db and querying for the flash amount. In the future this
can also be used for detecting 32MiB-RAM devices.

The list of packages that should be removed is defined in a variable
at the tob of the script.

In addition there is a list of routers that should get less packages
even if they have technically more than 8MiB flash. This is useful
for example for UniFi-AC-Mesh, which has two 8MiB-Partitions on a
16MiB-Flash.

Signed-off-by: Martin Hübner <martin.hubner@web.de>