build: explicitly define our branches
The freifunk_release file in falter-common is not the source of truth
for release branching and versioning anymore. The falter package feeds
are now rolling snapshot feeds just like OpenWrt's.

In this commit we stop reading the freifunk_release file and instead
explicitly define our release branching and versioning. It needs
to be updated whenever a new release branch is created, such as when
openwrt-22.03 was created and openwrt-19.07 removed.

In this commit we also start injecting a new freifunk_release file
into the image.

These two changes change how our releases are defined and started:

Previously we'd commit an updated freifunk_release file to the
falter-packages repo, then build the package feed, the build the
images.

Now all builds are the same and snapshot by default. We call
`build_falter` with `-v 1.2.3` for a snapshot release and with
`-v 1.2.3-snapshot` for a snapshot iamge. It takes care of all
branch-related and version-related information and injects it
into the image.

The differences b/w a snapshot image and release image are simple:
- Image filename
- FREIFUNK_RELEASE variable in /etc/freifunk_release
- Subdirectory on firmware.berlin.freifunk.net
  - stable/ vs. unstable/ (Buildbot will take care of this)
- Built using different schedulers (= buttons) in Buildbot

Things snapshot and release have in common:
- Built using the same Buildbot builder
- Built using OpenWrt snapshot imagebuilder (matching the branch)
- Using falter snapshot packages feed (matching the branch)

Missing pieces for the near future:
- FREIFUNK_REVISION in /etc/freifunk_release used to contain the
  falter package feed git revision, but we don't have knowledge of
  that during a targets build. OPENWRT_REVISION on the other hand
  contains the revision of openwrt.git, so maybe using the falter-builter.git
  revision would be better, anyway.