build: use last stable releases for 22.03 and 23.05
OpenWrt removes old kmods packages feeds of snapshots releases after a while.
A kind of garbage collection when that release branch is EOL.
The Imagebuilder can't build images if the kmods packages are unavailable.

This already happened in the past for 21.02-snapshot,
and now happened for 23.05-snapshot.
It'll probably happen for 22.03 as well soon.

Let's switch to the last stable patch release whenever a branch is EOL.
We only really want snapshots if fixes and patches still keep coming in.